refactor(GameRepository) : GameDto 추가 및 기능 분리
DTO 추가로 repository에서 데이터를 받는 방식 분리
