+++ src/main/java/racingcar/GameService.java
@@ -1,7 +1,6 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -9,39 +8,20 @@ public class GameService {
 
     GameRepository gameRepository = new GameRepository();
 
-
-    public void setCarsNames(String input) {
-
-        try {
-            inputValidator(input);
-            String delimiter = ",";
-            List<String> splitNames = List.of(input.split(delimiter));
-            nameLengthValidator(splitNames);
-            gameRepository.setCarName(splitNames);
-        } catch (Exception e) {
-            throw new IllegalArgumentException(e.getMessage());
-        }
-
-
+    public void gameNameSet(String input) {
+        gameRepository.setCarsNames(input);
     }
 
 
-    public void setNumberOfRounds(String input) {
-
-        try {
-            inputValidator(input);
-            int round = Integer.parseInt(input);
-            gameRepository.setRound(round);
-        } catch (Exception e) {
-            throw new IllegalArgumentException(e.getMessage());
-        }
-
+    public void gameRoundSet(String input) {
+        gameRepository.setNumberOfRounds(input);
     }
+
 
     public void playingGame() {
 
         Map<String, Integer> cars = setCarPositionOnTheLine();
-        for (int round = 0; round < gameRepository.getRound(); round++) {
+        for (int round = 0; round < gameRepository.roundOfGame(); round++) {
             roundOnRound(cars);
         }
         System.out.println("최종 우승자 : " + findingWinner(cars));
@@ -51,7 +31,7 @@ public class GameService {
 
     public Map<String, Integer> setCarPositionOnTheLine() {
 
-        return gameRepository.getCarNames().stream()
+        return gameRepository.getAllNames().stream()
                 .collect(Collectors.toMap(name -> name, name -> 0));
     }
 
@@ -95,25 +75,11 @@ public class GameService {
             Integer score = entry.getValue();
 
             if (score == maxValue) {
-                gameRepository.setWinner(car);
+                gameRepository.selectWinner(car);
             }
         }
 
-        return String.join(",", gameRepository.getWinner());
-    }
-
-    private void inputValidator(String input) {
-        if (input == null || input.isEmpty()) {
-            throw new IllegalArgumentException("입력값이 비어있습니다.");
-        }
-    }
-
-    private void nameLengthValidator(List<String> names) {
-        for (String name : names) {
-            if (name.length() > 5) {
-                throw new IllegalArgumentException("ERROR : 5자 이하의 이름만 입력 가능합니다.");
-            }
-        }
+        return String.join(",", gameRepository.findWinner());
     }
